Rename presenter module to repr.
