+++ src/_internal/rels.rs
@@ -2,7 +2,7 @@ use api::raw::{RawFetch, RawUpdate, Include};
 use api::{Resource, Entity, Result, Error};
 use api::rel::{Relation, LinkOne, UnlinkOne, ReplaceLinks, HasOne, HasMany, RelationId};
 use router::IncludeQuery;
-use presenter::Presenter;
+use repr::Presenter;
 
 pub trait _FetchRels: RawFetch {
     fn rels<P: Presenter>(entity: &Entity<Self>, includes: &[IncludeQuery]) -> Result<(Self::Relationships, Vec<Include<P>>)>;
