+++ src/api/mod.rs
@@ -1,7 +1,7 @@
 use std::result;
 use std::str::FromStr;
 
-use presenter::Represent;
+use repr::Represent;
 
 mod async;
 mod error;
